Do not redirect trailing slashes
Despite my best efforts, I have not found a way to redirect to
the correct hostname. It always redirects to the internal hostname,
which is obviously not helpful.

Instead, change APIs to not have a trailing slash and disable
automatic 307 redirects. This means a 404 will be returned for API
requests with a trailing slash.