Use dependency injection for determining access and querying database
This required moving from streaming results from the DB query to processing them
all and returning the result. This is due to how FastAPI resolves dependencies
internally.

It’s possible there is a way to use streaming results from the database with
dependency injection, but I do not currently know how.