Createt type refeneces for major and minor version
Sometimes they are path params, sometimes query params. Ideally these could be
defined once and reused, but I would have to create different versions for path
and query params, which gets really ugly.

It’s simpler to just defined them in the module they are used as the
correct type.