Combine module and package results in the relevant response
The end goal is have one entry in the respone per display name. For relevant
results that query inventory, the distinction between package and module isn’t
helpful, so that was removed from the response.

The fields used to compute the hash for the collection object were pared down to
achiev the desired response.