Improve RHEL version parsing from system profile
We ran into some data where operating_system[“name”] exists but both
major and minor keys are missing.

Instead of assuming that everything needed in operating_system is there
if the name field exists, test that we actually got a value for major. If not,
fall back to using os_release to get the major and minor versions.

Lastly, skip instead of failing if we don’t have the version information.