Use UUIDs from fixture data
Since the fixture data is being used to gather the sytem IDs, they need
to be set when loading.

The tests could get the system IDs by doing a database query instead of reading
the fixture file. Then we could go back to having the UUIDs generated each time
the data is loaded.