[RSPEED-1515] Change how package data are grouped (#206)
The goal of using a set initially was to keep the number package items that
needed to be processed to a minimum. But when system_id was added to the mix, it
remove any possibility of deduplicating across systems. Furthermore, that change
introduced a bug that overwrote instead of accumulating the package data while
processing the hosts.

Change the grouping strategy to collect system IDs into groups based on the
package and its major OS version.

Pare down the columns in the SQL query to only those that are needed. There is a lot
of data coming over the wire due to all the system data being stored in a single JSONB field.
This helps a little but there is more work to do with reducing the amount of data
returned from the database.

Yield in bigger batches. This helps when processing lots of hosts since it pulls from
the database in larger chunks.

Change the host data loader to create JSONB columns which matches the production schema.

Add tests to catch missing data regressions