Get uuids using 1 function and change set to list
Use one function for getting system UUIDs from the system_names list -
less duplicate code, easier to test.

Change set() in some classes to list() which is less memory efficient.
Unlikely allows duplicates - but this is handled in the database to not
contain two same hosts for one AppStream/System. Also hashing is not
needed, systems are only appended. The serialization of set and list is
the same in FastAPI.