Correct upcoming matching
Create a view of packages by host and use that when matching
upcoming items. This still needs a lot of work since the value we are
matching against is arbitrary data entered in an unstructed text field
in Jira. Based on the current values, people interpret that field to accept
a glab, a comma delimeted list of packages, and a proper name of project.
The matching right now explicitly compares the literal string values, which
will miss a whole lot of matches.