[RSPEED-1874] Correct matching for upcoming items (#231)
Create a view of packages by host and use that when matching
upcoming items. This still needs a lot of work since the value we are
matching against is arbitrary data entered in an unstructed text field
in Jira. Based on the current values, people interpret that field to accept
a glob, a comma delimited list of packages, and a proper name of project.
The matching right now explicitly compares the literal string values, which
will miss a whole lot of matches.

* Correct type hint
* Add OS major and minor to SystemInfo
* Correct upcoming matching
* Update system generation fixture
* Update tests based on new processing behavior
* Two more hosts match for nginx now
* Remove unused variable
* Make faker a test requirement
* Use a variable that doesn’t overlap with the function arg

* Remove unnecessary check for None
  The SystemInfo object requires os_major so this code is no longer necessary.