Correct relevant module return (#270)
* Change cache key a tuple instead of a string

The keys in the dictionary are tuples so this makes things more consistent.

* Only add the app stream to the cache if it has a start date

This matches the intention at the end of the function, which is to include the
object in the cache but not in the result. Currently we are incorrectly
returning app streams that did not accurately match.

* Check cache after evaluating module status

If the cache is evaluated before looking at the module status, a module may be
included in the results that should not be.

* Update tests